SquareOsc
had to include #include "pq_map_real.h"
#include "pq_time.h" in SquareOsc.cpp
